requirements.txt
# Core dependencies
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.2

# Deep learning
torch>=1.10.0
tensorflow>=2.8.0
transformers>=4.18.0
sentence-transformers>=2.2.0

# LLM API clients
openai>=1.0.0
anthropic>=0.4.0
cohere-api>=4.0.0
langchain>=0.0.200
langchain-community>=0.0.10

# Embeddings and vector databases
faiss-cpu>=1.7.0
chromadb>=0.4.0

# Image generation
diffusers>=0.14.0
stability-sdk>=0.8.0

# Web API
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=1.10.0

# Utilities
tqdm>=4.65.0
python-dotenv>=0.21.0
pytest>=7.3.0 